+++ NEWS.md
@@ -5,6 +5,7 @@ UNDER DEVELOPMENT
 **Implemented Updates**
 
 - `run_algos()`: add parameter to skip/warn/stop when trying to run DPA on non-DPA time series
+- Allow `run_algos()` to run on a list of `mds_ts` or other properly formatted objects
 
 **Bugfixes**
 
@@ -13,7 +14,6 @@ UNDER DEVELOPMENT
 
 **Potential Updates**
 
-- Allow `run_algos()` to run on a list of `mds_ts` objects
 - Add CUSUM
 
 `mdsstat` 0.1.0